moved Tooltip from Flamegraph to top level
