+++ packages/pyroscope-flamegraph/src/Tooltip/Tooltip.module.scss
